move auto-companies-and-contacts-creation folder
