+++ packages/twenty-server/src/engine/integrations/message-queue/jobs.module.ts
@@ -11,18 +11,18 @@ import { StripeModule } from 'src/engine/core-modules/billing/stripe/stripe.modu
 import { UserWorkspaceModule } from 'src/engine/core-modules/user-workspace/user-workspace.module';
 import { UserModule } from 'src/engine/core-modules/user/user.module';
 import { HandleWorkspaceMemberDeletedJob } from 'src/engine/core-modules/workspace/handle-workspace-member-deleted.job';
+import { WorkspaceModule } from 'src/engine/core-modules/workspace/workspace.module';
 import { EmailSenderJob } from 'src/engine/integrations/email/email-sender.job';
 import { EmailModule } from 'src/engine/integrations/email/email.module';
 import { DataSourceModule } from 'src/engine/metadata-modules/data-source/data-source.module';
 import { ObjectMetadataModule } from 'src/engine/metadata-modules/object-metadata/object-metadata.module';
 import { CleanInactiveWorkspaceJob } from 'src/engine/workspace-manager/workspace-cleaner/crons/clean-inactive-workspace.job';
-import { TimelineActivityModule } from 'src/modules/timeline/timeline-activity.module';
-import { WorkspaceModule } from 'src/engine/core-modules/workspace/workspace.module';
-import { AutoCompaniesAndContactsCreationJobModule } from 'src/modules/connected-account/auto-companies-and-contacts-creation/jobs/auto-companies-and-contacts-creation-job.module';
+import { AutoCompaniesAndContactsCreationJobModule } from 'src/modules/auto-companies-and-contacts-creation/jobs/auto-companies-and-contacts-creation-job.module';
+import { CalendarEventParticipantManagerModule } from 'src/modules/calendar/calendar-event-participant-manager/calendar-event-participant-manager.module';
+import { CalendarModule } from 'src/modules/calendar/calendar.module';
 import { MessagingModule } from 'src/modules/messaging/messaging.module';
 import { TimelineJobModule } from 'src/modules/timeline/jobs/timeline-job.module';
-import { CalendarModule } from 'src/modules/calendar/calendar.module';
-import { CalendarEventParticipantManagerModule } from 'src/modules/calendar/calendar-event-participant-manager/calendar-event-participant-manager.module';
+import { TimelineActivityModule } from 'src/modules/timeline/timeline-activity.module';
 
 @Module({
   imports: [
