+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/constants/contacts-creation-batch-size.constant.ts
