+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/create-company/create-company.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
 import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
+import { CreateCompanyService } from 'src/modules/auto-companies-and-contacts-creation/create-company/create-company.service';
 import { CompanyWorkspaceEntity } from 'src/modules/company/standard-objects/company.workspace-entity';
-import { CreateCompanyService } from 'src/modules/connected-account/auto-companies-and-contacts-creation/create-company/create-company.service';
 
 @Module({
   imports: [
