+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/create-company/create-company.service.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@nestjs/common';
 
+import axios, { AxiosInstance } from 'axios';
 import { EntityManager } from 'typeorm';
 import { v4 } from 'uuid';
-import axios, { AxiosInstance } from 'axios';
 
-import { CompanyRepository } from 'src/modules/company/repositories/company.repository';
 import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
+import { getCompanyNameFromDomainName } from 'src/modules/auto-companies-and-contacts-creation/utils/get-company-name-from-domain-name.util';
+import { CompanyRepository } from 'src/modules/company/repositories/company.repository';
 import { CompanyWorkspaceEntity } from 'src/modules/company/standard-objects/company.workspace-entity';
-import { getCompanyNameFromDomainName } from 'src/modules/connected-account/auto-companies-and-contacts-creation/utils/get-company-name-from-domain-name.util';
 @Injectable()
 export class CreateCompanyService {
   private readonly httpService: AxiosInstance;
