+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/types/contact.type.ts
