+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/utils/get-company-name-from-domain-name.util.ts
