+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/utils/get-domain-name-from-handle.util.ts
