+++ packages/twenty-server/src/modules/auto-companies-and-contacts-creation/utils/get-first-name-and-last-name-from-handle-and-display-name.util.ts
